📁  Refactor directory structure
