+++ src/common/Common.cpp
@@ -1,0 +1,3 @@
+//
+// Created by Study Together on 2024/02/11.
+//
